fix install jdk11

+++ scripts/install-jdk11.sh
@@ -1,21 +1,32 @@
 #!/bin/bash
 
-yum -y install tar wget
-mkdir -p /opt/java
-cd /opt/java
-wget http://PULSAR-CLIENT/pulsar/pkgs/jdk-11.0.15.1_linux-x64_bin.tar.gz
-tar zxvf jdk-11.0.15.1_linux-x64_bin.tar.gz
-rm jdk-11.0.15.1_linux-x64_bin.tar.gz
+yum -y install tar wget > ./yum.log 2>&1
 
+url="http://PULSAR-CLIENT/pulsar/pkgs/jdk-11.0.15.1_linux-x64_bin.tar.gz"
+if [[ $(grep -c "pulsar-distributed-cluster-jdk11" /etc/profile) -ne 0 ]]; then 
+	echo "[$(hostname)]exec> The JDK is already installed and will be skipped."; 
+else 
 
-cp /etc/profile /etc/profile.bak.beforeAddJDK11
+	mkdir -p /opt/java
+	cd /opt/java
+	
+	echo "[$(hostname)]exec> Download JDK from : $url"
+	wget -q http://PULSAR-CLIENT/pulsar/pkgs/jdk-11.0.15.1_linux-x64_bin.tar.gz
+	tar zxf jdk-11.0.15.1_linux-x64_bin.tar.gz
+	rm jdk-11.0.15.1_linux-x64_bin.tar.gz
+	
+	cp /etc/profile /etc/profile.bak.beforeAddJDK11
+	
+	echo "[$(hostname)]exec> Configure JDK env."
 
-echo "## JDK11 ##" | sudo tee -a /etc/profile
-echo 'export JAVA_HOME=/opt/java/jdk-11.0.15.1' | sudo tee -a /etc/profile
-echo 'export JRE_HOME=$JAVA_HOME/jre' | sudo tee -a /etc/profile
-echo 'export CLASSPATH=$CLASSPATH:$JAVA_HOME/lib:$JAVA_HOME/jre/lib' | sudo tee -a /etc/profile
-echo 'export PATH=$PATH:$JAVA_HOME/bin:$JAVA_HOME/jre/bin:$HOME/bin' | sudo tee -a /etc/profile
-echo "## JDK11 ##" | sudo tee -a /etc/profile
+	echo "## pulsar-distributed-cluster-jdk11 ##" | sudo tee -a /etc/profile
+	echo 'export JAVA_HOME=/opt/java/jdk-11.0.15.1' | sudo tee -a /etc/profile
+	echo 'export JRE_HOME=$JAVA_HOME/jre' | sudo tee -a /etc/profile
+	echo 'export CLASSPATH=$CLASSPATH:$JAVA_HOME/lib:$JAVA_HOME/jre/lib' | sudo tee -a /etc/profile
+	echo 'export PATH=$PATH:$JAVA_HOME/bin:$JAVA_HOME/jre/bin:$HOME/bin' | sudo tee -a /etc/profile
+	echo "## pulsar-distributed-cluster-jdk11 ##" | sudo tee -a /etc/profile
+
+fi
 
 source /etc/profile
 java -version
